refactor: 경로 수정 및 알파벳 컨벤션 통일
