+++ src/helpers/OutputManager.js
@@ -18,13 +18,13 @@ class OutputManager {
     });
   }
 
-  static printLottoResult(LottoResult) {
+  static printLottoResult(lottoResult) {
     this.print(OUTPUT_MESSAGE.WINNING_STATICS);
 
     Object.entries(LOTTO_RANK_MAP).forEach(([rawRank, lottoInfo]) => {
       const rank = +rawRank;
       const { winningCount, prizeMoney } = lottoInfo;
-      const winningLottoCount = LottoResult.getWinningCount(rank);
+      const winningLottoCount = lottoResult.getWinningCount(rank);
 
       const bonusNumberString = this.#getBonusNumberString(rank);
       const winningLottoOutput = `${winningCount}개 일치${bonusNumberString} (${prizeMoney.toLocaleString(
