+++ src/objects/index.js
@@ -2,3 +2,4 @@ export { default as Lotto } from './Lotto.js';
 export { default as LottoBuyer } from './LottoBuyer.js';
 export { default as LottoCompany } from './LottoCompany.js';
 export { default as LottoShop } from './LottoShop.js';
+export { default as LottoResult } from './LottoResult.js';
